put Today>Task functionality in useTask
